Add PROBE/SEQUENCE to MFMC only if new addition will be unique
Helper functions to add PROBE or SEQUENCE to HDF5 file if candidate is unique.

E.g. If probe to be added is already present in file, then no addition is made, but index/ref/location relating to original is returned instead. Mandatory entries (coordinates, centre frequency etc) are checked to determine uniqueness.

With regards sequence, the TX/RX laws are compared along with start time and time step.